scripts/provision/requirements.txt
# Provisioning Script Dependencies

# AWS SDK
boto3>=1.28.0
botocore>=1.31.0

# Configuration file handling
pyyaml>=6.0

# CLI argument parsing (built-in)
# argparse

# File operations (built-in)
# pathlib
# json
# shutil
